Big Patch
Changed the files so that the .wav are written and stored easily. Also added the variables so that changes to the names do not need to be made one by one.
Written the final stereo but needs more looking into it
Changed the noise addition so it¡s even quietter